+++ packages/components/src/components/dropdown/dropdown.styles.ts
@@ -1,8 +1,6 @@
-/* eslint-disable */
 import { css } from 'lit';
 
 export default css`
-	/* stylelint-disable */
   :host {
     display: inline-block;
   }
@@ -32,11 +30,11 @@ export default css`
   }
 
   .dropdown__panel {
+    border-radius: var(--syn-border-radius-medium);
+    box-shadow: var(--syn-shadow-large);
     font-family: var(--syn-font-sans);
     font-size: var(--syn-font-size-medium);
     font-weight: var(--syn-font-weight-normal);
-    box-shadow: var(--syn-shadow-large);
-    border-radius: var(--syn-border-radius-medium);
     pointer-events: none;
   }
 
@@ -47,7 +45,7 @@ export default css`
 
   /* When users slot a menu, make sure it conforms to the popup's auto-size */
   ::slotted(syn-menu) {
-    max-width: var(--auto-size-available-width) !important;
     max-height: var(--auto-size-available-height) !important;
+    max-width: var(--auto-size-available-width) !important;
   }
 `;