+++ packages/components/src/components/switch/switch.component.ts
@@ -10,10 +10,8 @@ import { property, query, state } from 'lit/decorators.js';
 import { watch } from '../../internal/watch.js';
 import componentStyles from '../../styles/component.styles.js';
 import formControlStyles from '../../styles/form-control.styles.js';
-import formControlCustomStyles from '../../styles/form-control.custom.styles.js';
 import SynergyElement from '../../internal/synergy-element.js';
 import styles from './switch.styles.js';
-import customStyles from './switch.custom.styles.js';
 import type { CSSResultGroup } from 'lit';
 import type { SynergyFormControl } from '../../internal/synergy-element.js';
 import { enableDefaultSettings } from '../../utilities/defaultSettings/decorator.js';
@@ -45,7 +43,7 @@ import { enableDefaultSettings } from '../../utilities/defaultSettings/decorator
  */
 @enableDefaultSettings('SynSwitch')
 export default class SynSwitch extends SynergyElement implements SynergyFormControl {
-  static styles: CSSResultGroup = [componentStyles, formControlStyles, styles, formControlCustomStyles, customStyles];
+  static styles: CSSResultGroup = [componentStyles, formControlStyles, styles];
 
   private readonly formControlController = new FormControlController(this, {
     value: (control: SynSwitch) => (control.checked ? control.value || 'on' : undefined),
@@ -70,6 +68,9 @@ export default class SynSwitch extends SynergyElement implements SynergyFormCont
 
   /** Disables the switch. */
   @property({ type: Boolean, reflect: true }) disabled = false;
+
+  /** Sets the switch to a readonly state. */
+  @property({ type: Boolean, reflect: true }) readonly = false;
 
   /** Draws the switch in a checked state. */
   @property({ type: Boolean, reflect: true }) checked = false;
@@ -118,7 +119,12 @@ export default class SynSwitch extends SynergyElement implements SynergyFormCont
     this.formControlController.emitInvalidEvent(event);
   }
 
-  private handleClick() {
+  private handleClick(e: MouseEvent) {
+    if (this.readonly) {
+      e.preventDefault();
+      return;
+    }
+
     this.checked = !this.checked;
     this.emit('syn-change');
   }
@@ -129,6 +135,10 @@ export default class SynSwitch extends SynergyElement implements SynergyFormCont
   }
 
   private handleKeyDown(event: KeyboardEvent) {
+    if (this.readonly) {
+      return;
+    }
+
     if (event.key === 'ArrowLeft') {
       event.preventDefault();
       this.checked = false;
@@ -212,6 +222,7 @@ export default class SynSwitch extends SynergyElement implements SynergyFormCont
             switch: true,
             'switch--checked': this.checked,
             'switch--disabled': this.disabled,
+            'switch--readonly': this.readonly,
             'switch--focused': this.hasFocus,
             'switch--small': this.size === 'small',
             'switch--medium': this.size === 'medium',
@@ -226,6 +237,7 @@ export default class SynSwitch extends SynergyElement implements SynergyFormCont
             value=${ifDefined(this.value)}
             .checked=${live(this.checked)}
             .disabled=${this.disabled}
+            .readOnly=${this.readonly}
             .required=${this.required}
             role="switch"
             aria-checked=${this.checked ? 'true' : 'false'}